SAP-iXp-Sftw-Dev-Project-Examples/camera-recognition-main/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
pytesseract>=0.3.8
